requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
blis==1.0.1
catalogue==2.0.10
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudpathlib==0.20.0
confection==0.1.5
cymem==2.0.8
da_core_news_sm @ https://github.com/explosion/spacy-models/releases/download/da_core_news_sm-3.8.0/da_core_news_sm-3.8.0-py3-none-any.whl#sha256=8caeb4dc26f56de8abcf70399b202a0b2223d38664b1a59f1cb8db0000808bc9
Django==4.2.16
filelock==3.16.1
fsspec==2024.10.0
gunicorn==23.0.0
huggingface-hub==0.26.2
idna==3.10
Jinja2==3.1.4
langcodes==3.4.1
language_data==1.2.0
lxml==5.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.10
networkx==3.2.1
numpy==2.0.2
packaging==24.2
pipdeptree==2.23.4
preshed==3.0.9
protobuf==5.28.3
psycopg2==2.9.10
punctfix==0.11.1
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
python-docx==1.1.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
safetensors==0.4.5
sentencepiece==0.2.0
shellingham==1.5.4
smart-open==7.0.5
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.2
srsly==2.4.8
sympy==1.13.1
thinc==8.3.2
tokenizers==0.20.3
torch==2.5.1
tqdm==4.67.0
transformers==4.46.2
typer==0.13.1
typing_extensions==4.12.2
urllib3==2.2.3
wasabi==1.1.3
weasel==0.4.1
wrapt==1.16.0
